requirements.txt
# Core Framework & LLM
agno>=0.1.0
anthropic>=0.7.0
python-dotenv>=0.19.0

# Web Framework
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
websockets>=11.0

# XML & Document Processing
defusedxml>=0.0.1
lxml>=4.9.0
python-docx>=0.8.11

sqlalchemy
